LAB 10 - Updating data from application to server Add / remove / modify movies from our own application Notes: we could've handled the aftermath of sending a request as showing some kind of loading screen, or intermediate state of the application. This could've easily been done by adding an intermediate state with 'loading' or 'processing' => true / false, and when the response was received, if positive, update the value to its counterpart and update the entire state of the application by refreshing the page or by changing a value on which useEffect() in the main application page (which contains the data from the movies database) was contained.
Except for this, all that was new was simply sending requests to the server through fetch() specifying request options such as headers, body and application type.
Also, fetch returns a different object than a normal promise, so you have to adapt to its methods.

This means that for example, when a fetch() request is succesful, it will return a status of 'ok' - this doesn't meant that it contains a status of 200,
it just means that the request response has been succesfully delivered.
It is our responsability to check its content and its status / data - this is done in this lab for example by checking the type of content inside of the
response.headers. (of course, this is handled by us as we designed the API and know that if the request is succesful, the response will contain a json object so we can check for an 'application/json' type content.)